Small Update to Error Checking.
Now detects invalid arc files that are not in the inverse endian.
Stops ArcFill process if LoadArc function returns null, stopping the program from crashing when trying to open an invalid or unsupported Arc.
FlushAnndClose function now detects if variable is null to fix crash if Close is selected without opening a file.